test(db): cover pg.ts — run the store contract against real Postgres
The db coverage left one real gap: pg.ts (the hosted-tier Postgres driver,
~34KB of raw SQL mirroring repos.ts) was at 0% measured. PR #127 excluded it,
reasoning the api suite exercises it under DOCK_TEST_DB=pg — true, but
unmeasured, and the db package's own comprehensive round-trip only drove SQLite,
so pg-only methods (claimDueDeliveries' FOR UPDATE SKIP LOCKED lease, domain
uniqueness, agent mention ack, moderation, notifications) had no direct test.

Extract the 12 entity-lifecycle suites into a shared store contract
(store-contract.ts) and run it against both dialects:
- sqlite-store.test.ts: in-memory SQLite, every `pnpm test` (zero-config)
- pg-store.test.ts: a real Postgres when DOCK_TEST_DB=pg, each run in its own
  search_path schema (created + dropped per run); a no-op otherwise

The SAME assertions must hold on both drivers, so a wrong WHERE / missing org
scope / broken transaction in pg.ts fails the assertion that passes on SQLite.
Verified against a live Postgres: pg.ts 0% -> 93.88% lines / 90% stmts.

Wire it into the existing `pg` lane: scripts/test-pg.sh runs the db contract
after the api suite, and a scoped gate (vitest.pg.config.ts, 88/85/55) measures
+ ratchets pg.ts there. The default sqlite gate is unchanged. d1.ts stays out
(the Cloudflare-D1 driver needs the workerd runtime).
