+++ packages/db/package.json
@@ -20,6 +20,7 @@
     "typecheck": "tsc --noEmit",
     "test": "vitest run --passWithNoTests",
     "test:coverage": "vitest run --coverage",
+    "test:pg": "vitest run --config vitest.pg.config.ts --coverage test/pg-store.test.ts",
     "gen:d1-schema": "vitest run test/d1-schema.test.ts -u"
   },
   "dependencies": {
